+++ r4/src/test/java/io/github/jy95/fds/r4/translators/TimeOfDayTest.java
@@ -1,82 +1,53 @@
 package io.github.jy95.fds.r4.translators;
 
-import io.github.jy95.fds.r4.DosageAPIR4;
-import io.github.jy95.fds.r4.AbstractFhirTest;
 import io.github.jy95.fds.common.types.DisplayOrder;
+import io.github.jy95.fds.common.types.DosageAPI;
+import io.github.jy95.fds.r4.DosageAPIR4;
+import io.github.jy95.fds.r4.config.FDSConfigR4;
+import io.github.jy95.fds.translators.AbstractTimeOfDayTest;
 import org.hl7.fhir.r4.model.Dosage;
 import org.hl7.fhir.r4.model.Timing;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
 
+import java.util.List;
 import java.util.Locale;
-import java.util.concurrent.ExecutionException;
 
-import static org.junit.jupiter.api.Assertions.*;
+public class TimeOfDayTest extends AbstractTimeOfDayTest<FDSConfigR4, Dosage> {
 
-public class TimeOfDayTest extends AbstractFhirTest {
-
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testNoTimeOfDay(Locale locale) throws ExecutionException, InterruptedException {
-        Dosage dosage = new Dosage();
-        DosageAPIR4 dosageUtils = TimeOfDayTest.getDosageAPI(locale, DisplayOrder.TIME_OF_DAY);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        assertEquals("", result);
-    }
-
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testSingleTimeOfDay(Locale locale) throws ExecutionException, InterruptedException {
+    @Override
+    protected Dosage generateSingleTimeOfDay() {
         Dosage dosage = new Dosage();
         Timing timing = new Timing();
         timing.getRepeat().addTimeOfDay("15:00:00");
         dosage.setTiming(timing);
-
-        DosageAPIR4 dosageUtils = TimeOfDayTest.getDosageAPI(locale, DisplayOrder.TIME_OF_DAY);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expectedResult = getExpectedSingleTimeText(locale);
-        assertEquals(expectedResult, result);
+        return dosage;
     }
 
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testMultipleTimeOfDay(Locale locale) throws ExecutionException, InterruptedException {
+    @Override
+    protected Dosage generateMultipleTimeOfDay() {
         Dosage dosage = new Dosage();
         Timing timing = new Timing();
         timing.getRepeat().addTimeOfDay("15:00:00");
         timing.getRepeat().addTimeOfDay("12:12:12");
         dosage.setTiming(timing);
-
-        DosageAPIR4 dosageUtils = TimeOfDayTest.getDosageAPI(locale, DisplayOrder.TIME_OF_DAY);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expectedResult = getExpectedMultipleTimesText(locale);
-        assertEquals(expectedResult, result);
+        return dosage;
     }
 
-    // Expected text for single timeOfDay
-    private static String getExpectedSingleTimeText(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "at 15:00";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "à 15:00";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "um 15:00";
-        } else {
-            return "om 15:00";
-        }
+    @Override
+    public DosageAPI<FDSConfigR4, Dosage> getDosageAPI(Locale locale, DisplayOrder displayOrder) {
+        return new DosageAPIR4(FDSConfigR4.builder()
+                .displayOrder(List.of(displayOrder))
+                .locale(locale)
+                .build());
     }
 
-    // Expected text for multiple timeOfDay
-    private static String getExpectedMultipleTimesText(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "at 15:00 and 12:12:12";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "à 15:00 et 12:12:12";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "um 15:00 und 12:12:12";
-        } else {
-            return "om 15:00 en 12:12:12";
-        }
+    @Override
+    public DosageAPI<FDSConfigR4, Dosage> getDosageAPI(FDSConfigR4 config) {
+        return new DosageAPIR4(config);
+    }
+
+    @Override
+    public Dosage generateEmptyDosage() {
+        return new Dosage();
     }
 
 }
